st_courses_1/AAD/requirements.txt
mysql-connector-python==9.0.0
# mysqlclient==2.2.4
SQLAlchemy==1.4.23
joblib==1.3.2
scikit-learn==1.3.2
numpy==1.26.4
pandas==2.1.2
openpyxl>=3.1.2
